PIPE-10455: Fix Type Casting and Overflow Issue During Multiplication (#7)
* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* Fix integer overflow in X86CallLowering.cpp line 65 - cast operand before multiplication

* Fix 13 integer overflow vulnerabilities in X86 backend

- X86CallLowering.cpp line 65: Cast uint to uint64_t before multiplication
- X86ISelLowering.cpp (12 lines): Cast operands before multiplication
  - Lines: 4699, 7843, 9667, 18143, 18145, 26499, 29168, 32156, 32172, 32279, 32411, 32952
  - Pattern: Use (size_t) for array operations, (uint64_t) for DAG constants

Security fixes for CWE-190, CWE-192, CWE-197, CWE-681

* Fix Printf format specifiers in FuzzerCorpus.h - use %zu for size_t, %u for uint32_t (CWE-686)

* Fix signed integer overflow in IndVarSimplify.cpp - use int64_t for safe overflow checks (CWE-190, CWE-128)

* Fix multiplication overflow in ARMLoadStoreOptimizer.cpp - cast WordOffset to int64_t before multiplication (CWE-190/192/197/681)

* Potential fix for pull request finding 'CodeQL / Wrong type of arguments to formatting function'


* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* Update TargetLowering.cpp

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* PIPE-10455: SAST findings of type Multiplication result converted to larger type

* Update raw_ostream.h

* Update MachODump.cpp

* Update PPCISelLowering.cpp

---------

+++ lib/CodeGen/MachineTraceMetrics.cpp
@@ -71,7 +71,7 @@ bool MachineTraceMetrics::runOnMachineFunction(MachineFunction &Func) {
   Loops = &getAnalysis<MachineLoopInfo>();
   SchedModel.init(ST.getSchedModel(), &ST, TII);
   BlockInfo.resize(MF->getNumBlockIDs());
-  ProcResourceCycles.resize(MF->getNumBlockIDs() *
+  ProcResourceCycles.resize((uint64_t)MF->getNumBlockIDs() *
                             SchedModel.getNumProcResourceKinds());
   return false;
 }
@@ -145,8 +145,8 @@ MachineTraceMetrics::getProcResourceCycles(unsigned MBBNum) const {
   assert(BlockInfo[MBBNum].hasResources() &&
          "getResources() must be called before getProcResourceCycles()");
   unsigned PRKinds = SchedModel.getNumProcResourceKinds();
-  assert((MBBNum+1) * PRKinds <= ProcResourceCycles.size());
-  return makeArrayRef(ProcResourceCycles.data() + MBBNum * PRKinds, PRKinds);
+  assert((uint64_t)(MBBNum+1) * PRKinds <= ProcResourceCycles.size());
+  return makeArrayRef(ProcResourceCycles.data() + (uint64_t)MBBNum * PRKinds, PRKinds);
 }
 
 //===----------------------------------------------------------------------===//
@@ -264,8 +264,8 @@ ArrayRef<unsigned>
 MachineTraceMetrics::Ensemble::
 getProcResourceDepths(unsigned MBBNum) const {
   unsigned PRKinds = MTM.SchedModel.getNumProcResourceKinds();
-  assert((MBBNum+1) * PRKinds <= ProcResourceDepths.size());
-  return makeArrayRef(ProcResourceDepths.data() + MBBNum * PRKinds, PRKinds);
+  assert((uint64_t)(MBBNum+1) * PRKinds <= ProcResourceDepths.size());
+  return makeArrayRef(ProcResourceDepths.data() + (uint64_t)MBBNum * PRKinds, PRKinds);
 }
 
 /// Get an array of processor resource heights for MBB. Indexed by processor
@@ -277,8 +277,8 @@ ArrayRef<unsigned>
 MachineTraceMetrics::Ensemble::
 getProcResourceHeights(unsigned MBBNum) const {
   unsigned PRKinds = MTM.SchedModel.getNumProcResourceKinds();
-  assert((MBBNum+1) * PRKinds <= ProcResourceHeights.size());
-  return makeArrayRef(ProcResourceHeights.data() + MBBNum * PRKinds, PRKinds);
+  assert((uint64_t)(MBBNum+1) * PRKinds <= ProcResourceHeights.size());
+  return makeArrayRef(ProcResourceHeights.data() + (uint64_t)MBBNum * PRKinds, PRKinds);
 }
 
 //===----------------------------------------------------------------------===//
